refactor: narrow A1Z integration scope
