--- dimos/learning/dataprep/galaxea_a1z_camera_config.json
@@ -1,37 +1,0 @@
-{
-  "source": "",
-  "episodes": {
-    "extractor": "episode_status",
-    "status_stream": "status"
-  },
-  "observation": {
-    "image": {
-      "stream": "color_image",
-      "field": "data"
-    },
-    "joint_state": {
-      "stream": "coordinator_joint_state",
-      "field": "position"
-    }
-  },
-  "action": {
-    "joint_target": {
-      "stream": "coordinator_joint_state",
-      "field": "position"
-    }
-  },
-  "sync": {
-    "anchor": "image",
-    "rate_hz": 30.0,
-    "tolerance_ms": 50.0,
-    "action_shift": 1
-  },
-  "output": {
-    "format": "lerobot",
-    "path": "data/datasets/galaxea_a1z_camera",
-    "metadata": {
-      "robot": "galaxea_a1z",
-      "default_task_label": "hand_teach"
-    }
-  }
-}
